fix[Store] : Fix compilation warning in file storage

+++ mooncake-store/src/file_storage.cpp
@@ -525,7 +525,7 @@ tl::expected<void, ErrorCode> FileStorage::GroupOffloadingKeysByBucket(
         }
 
         // Fill the rest of the bucket with new offloading objects
-        for (size_t i = bucket_keys.size(); i < config_.bucket_keys_limit;
+        for (int64_t i = static_cast<int64_t>(bucket_keys.size()); i < config_.bucket_keys_limit;
              ++i) {
             if (it == offloading_objects.cend()) {
                 // No more objects to add — move current batch to ungrouped pool
